support urls with non-root extensions
the current implementation does not support urls such as example.com/typesnse
or example.com/search
due to the way the uriFor constructs with the port.

This is now supported by checking if an extension is present, and
injecting the port at the proper location

example.com:443
example.com:443/typesense
etc.