feat: add type for optional array search params
Introduces a new utility type, ArrayableSearchParams, and a constant,
searchParamsArrayKeys. The ArrayableSearchParams typefilters the
properties of the SearchParams type, selecting only thosewhose type is
an array. The searchParamsArrayKeys constant is anobject that indicates
which parameters are expected to be arrays. Thisenhancement provides a
compile-time check for array parameters,improving type safety and code
readability.